// packages/data-atlas-wizard/src/generated/EndpointKind.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

/**
 * EndpointKind
 * @generated
 */
export const EndpointKind = {
  REST: 'REST',
  GEOJSON: 'GEOJSON',
  XMLA: 'XMLA',
  GRAPHQL: 'GRAPHQL',
  QGIS: 'QGIS',
  OGC_FEATURES: 'OGC_FEATURES',
  OGC_SENSORTHINGS: 'OGC_SENSORTHINGS',
  ODATA: 'ODATA'
} as const;

export type EndpointKind = typeof EndpointKind[keyof typeof EndpointKind];

/**
 * Get enum value by name
 */
export function getEndpointKind(name: string): EndpointKind | undefined {
  return (EndpointKind as Record<string, string>)[name] as EndpointKind | undefined;
}

/**
 * Get enum value by literal
 */
export function getEndpointKindByLiteral(literal: string): EndpointKind | undefined {
  if (literal === 'REST') return EndpointKind.REST;
  if (literal === 'GEOJSON') return EndpointKind.GEOJSON;
  if (literal === 'XMLA') return EndpointKind.XMLA;
  if (literal === 'GRAPHQL') return EndpointKind.GRAPHQL;
  if (literal === 'QGIS') return EndpointKind.QGIS;
  if (literal === 'OGC_FEATURES') return EndpointKind.OGC_FEATURES;
  if (literal === 'OGC_SENSORTHINGS') return EndpointKind.OGC_SENSORTHINGS;
  if (literal === 'ODATA') return EndpointKind.ODATA;
  return undefined;
}