// packages/data-atlas-wizard/src/generated/EndpointConfig.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';
import type { EndpointKind } from './EndpointKind';
import type { DataChain } from './DataChain';
import type { EndpointEntry } from './EndpointEntry';

/**
 * EndpointConfig
 * @generated
 */
export interface EndpointConfig extends EObject {
  id: string;
  name: string;
  description: string;
  kind: EndpointKind;
  urlContext: string;
  chains: DataChain[];
  openApi?: boolean;
  paginationOffsetParameterName?: string;
  paginationSizeParameterName?: string;
  entries: EndpointEntry[];

}